obtain and compare custom univariate summary statistics from given two datasets

This function could obtain and compare statistical information from given two datasets.

ddiff_info_univariate(
  dat1,
  dat2,
  key,
  measure_arg_con = list(min = min, max = max, med = median, mean = mean, std =
    "standard error"),
  measure_arg_cat = list(table = table),
  measure_arg_bin = list(table = table),
  dis_arg = list("euclidean", "manhattan", "gower"),
  ...
)

Arguments

dat1
First Data need to be compared.

dat2
Second Data need to be compared.

key
Unique identifiers to link two datasets

measure_arg_con
list of function for summary statistics of continuous response. The defalut list is c("min", "max", "median", "mean", "standard error")

measure_arg_cat
list of function for summary statistics of categorical response

measure_arg_bin
list of function for summary statistics of binary response

dis_arg
list of distance for measuring the difference of variables in each datasets

...
potential argument to added

Value

TBC

Examples

d <- test_data()
ddiff_rpt <- ddiff_info_univariate(d$identical$new, d$identical$old, c("ID", "Days"))
ddiff_rpt

#> $result_con
#>              stat    variable         data1         data2
#> 1             min        Days  1.000000e+00  1.000000e+00
#> 2             min          ID  1.000000e+00  1.000000e+00
#> 3             min       nEggs  0.000000e+00  0.000000e+00
#> 4             min nEggsRemain  0.000000e+00  0.000000e+00
#> 5             min        nm_1 -4.196442e+00 -4.196442e+00
#> 6             min        nm_2 -3.479414e+00 -3.479414e+00
#> 7             min        nm_3 -3.580228e+00 -3.580228e+00
#> 8             max        Days  2.500000e+01  2.500000e+01
#> 9             max          ID  1.000000e+03  1.000000e+03
#> 10            max       nEggs  1.400000e+02  1.400000e+02
#> 11            max nEggsRemain  1.757000e+03  1.757000e+03
#> 12            max        nm_1  3.992569e+00  3.992569e+00
#> 13            max        nm_2  4.006142e+00  4.006142e+00
#> 14            max        nm_3  4.118953e+00  4.118953e+00
#> 15            med        Days  1.300000e+01  1.300000e+01
#> 16            med          ID  4.800000e+02  4.800000e+02
#> 17            med       nEggs  1.600000e+01  1.600000e+01
#> 18            med nEggsRemain  2.230000e+02  2.230000e+02
#> 19            med        nm_1 -5.383370e-03 -5.383370e-03
#> 20            med        nm_2  3.603452e-03  3.603452e-03
#> 21            med        nm_3  1.597123e-02  1.597123e-02
#> 22           mean        Days  1.300000e+01  1.300000e+01
#> 23           mean          ID  4.757250e+02  4.757250e+02
#> 24           mean       nEggs  2.330910e+01  2.330910e+01
#> 25           mean nEggsRemain  2.953232e+02  2.953232e+02
#> 26           mean        nm_1 -3.207803e-03 -3.207803e-03
#> 27           mean        nm_2  3.202436e-03  3.202436e-03
#> 28           mean        nm_3  9.157477e-03  9.157477e-03
#> 29 standard error        Days  5.134571e-02  5.134571e-02
#> 30 standard error          ID  1.987285e+00  1.987285e+00
#> 31 standard error       nEggs  1.816081e-01  1.816081e-01
#> 32 standard error nEggsRemain  2.103451e+00  2.103451e+00
#> 33 standard error        nm_1  7.064353e-03  7.064353e-03
#> 34 standard error        nm_2  7.115586e-03  7.115586e-03
#> 35 standard error        nm_3  7.111697e-03  7.111697e-03
#> 
#> $result_cat
#> $result_cat[[1]]
#>   cat_1 data1 data2
#> 1     a  6674  6674
#> 2     b  6526  6526
#> 3     c  6525  6525
#> 
#> $result_cat[[2]]
#>   cat_2 data1 data2
#> 1     d  6642  6642
#> 2     e  6429  6429
#> 3     f  6654  6654
#> 
#> 
#> $result_bin
#> $result_bin[[1]]
#>   bin_1 data1 data2
#> 1     N  9876  9876
#> 2     Y  9849  9849
#> 
#> $result_bin[[2]]
#>   bin_2 data1 data2
#> 1     F  9859  9859
#> 2     M  9866  9866
#> 
#>