compare given two datasets in form level

This function could obtain and compare empirical information of row, column and attributes from given two datasets. The comparison are made on following aspects :

whether the types of variables from given two datasets are the same

whether column names from given two datasets are the same

whether row names from given two datasets are the same

whether attributes from given two datasets are the same

whether the information given two datasets are the same (shuffled rows or columns)

whether the given two datasets are equivalent in R memory

ddiff_form(dat1, dat2, key, ...)

Arguments

dat1

First Data need to be compared.

dat2

Second Data need to be compared.

key

Unique identifiers to link two datasets

...

potential argument to added

Value

ddiff which contains indicator of above aspects. Also, the last element include existing comparedf object.

Examples

d <- test_data()
ddiff_rpt <- ddiff_form(d$identical$new, d$identical$old, c("ID", "Days"))
ddiff_rpt
#> $identical
#> [1] TRUE
#> 
#> $attributes_match
#> [1] TRUE
#> 
#> $column_match
#> [1] TRUE
#> 
#> $row_match
#> [1] TRUE
#> 
#> $info_match
#> [1] TRUE
#> 
#> $message
#> [1] "data1 and data2 are identical"
#> 
#> $comparedf
#> Compare Object
#> 
#> Function Call: 
#> arsenal::comparedf(x = dat1, y = dat2)
#> 
#> Shared: 12 non-by variables and 19725 observations.
#> Not shared: 0 variables and 0 observations.
#> 
#> Differences found in 0/12 variables compared.
#> 0 variables compared have non-identical attributes.
#>